agents/requirements.txt
flask
flask-cors
python-dotenv
uagents
fetchai
google-generativeai
requests